+++ packages/alfa-rules/src/common/predicate/text-spacing.ts
