+++ packages/alfa-rules/tsconfig.json
@@ -7,7 +7,7 @@
     "src/common/expectation.ts",
     "src/common/expectation/media-text-alternative.ts",
     "src/common/expectation/outcome-to-trilean.ts",
-    "src/common/expectation/text-spacing.ts",
+    "src/common/predicate/text-spacing.ts",
     "src/common/expectation/video-description-track-accurate.ts",
     "src/common/group.ts",
     "src/common/outcome/text-spacing.ts",
